backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
alembic==1.11.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
python-dotenv==1.0.0
slowapi==0.1.5

# For password hashing
bcrypt

# For UUID generation if used (though SQLModel handles UUIDs often)
# uuid # already in Python standard lib `uuid.uuid4()`

# If using async pg driver like asyncpg:
# asyncpg

# Testing
pytest==7.4.2
pytest-asyncio==0.22.0
anyio==3.7.1